Update SVD to workaround broken fields
Some SVD files define declare more than one fields,
whereas the bitWidth is greater than 1.
This is not practical and worse, in the datasheet, they are designed
as one bitfield, e.g A instead of A1, A2 and A3.
This adds a way to convert the broken bitfield to one field.

Signed-off-by: Alexandre Bailon <abailon@baylibre.com>